refactor: MovieListActivity -> MovieListFragment 수정
