+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -7,7 +7,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.movielist.view.MovieListActivity">
+        tools:context=".ui.movielist.view.MovieListFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/movies_recycler_view"
